pyproject.toml
[tool.poetry]
name = "usa-jobs-data"
version = "0.1.0"
description = ""
authors = ["Betania Martinez <martinezbetania@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.32.3"
numpy = "1.26.4"
redshift-connector = "^2.1.3"
sqlalchemy-redshift = "^0.8.14"
psycopg2 = "^2.9.9"
python-dotenv = "1.0.0"
pandas = "^2.2.3"
pyarrow = "^17.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

